fix(PIO-52): extract StegoForm component from SecretForm pattern for consistent styling
Revert TextAreaInput changes. Create StegoForm.vue using the same structure
and textarea usage as SecretForm.vue so visual styling is naturally identical.
StegoPage.vue is now a thin wrapper that renders StegoForm.
